+++ variables/discrete/PoissonRand.h
@@ -19,7 +19,7 @@ public:
 
     virtual double P(int k) const override;
     virtual double F(double x) const override;
-    virtual int variate() override;
+    virtual double variate() override;
 
     double E() const override { return l; }
     double Var() const override { return l; }
